+++ bee-block/src/protocol.rs
@@ -99,7 +99,7 @@ mod inx {
         type Error = crate::error::inx::InxError;
 
         fn try_from(value: inx_bindings::proto::RawProtocolParameters) -> Result<Self, Self::Error> {
-            Self::unpack_verified(value.params).map_err(|e| InxError::InvalidRawBytes(format!("{:?}", e)))
+            Self::unpack_verified(value.params, &()).map_err(|e| InxError::InvalidRawBytes(format!("{:?}", e)))
         }
     }
 }