+++ bee-block/tests/treasury_transaction_payload.rs
@@ -43,7 +43,7 @@ fn pack_unpack_valid() {
     )
     .unwrap();
     let transaction_2 =
-        TreasuryTransactionPayload::unpack_verified(&mut transaction_1.pack_to_vec().as_slice()).unwrap();
+        TreasuryTransactionPayload::unpack_verified(&mut transaction_1.pack_to_vec().as_slice(), &()).unwrap();
 
     assert_eq!(transaction_1, transaction_2);
 }
@@ -58,5 +58,5 @@ fn pack_unpack_invalid() {
 
     // Actual error is not checked because the output type check is done after the output is parsed so the error is not
     // `InvalidOutputKind` but something related to an invalid address, so not really relevant for this test.
-    assert!(TreasuryTransactionPayload::unpack_verified(&mut bytes.as_slice()).is_err());
+    assert!(TreasuryTransactionPayload::unpack_verified(&mut bytes.as_slice(), &()).is_err());
 }