+++ bee-ledger/bee-ledger-types/CHANGELOG.md
@@ -11,8 +11,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Bump `bee-block` to `v1.0.0-beta.4`;
-
 ### Deprecated
 
 ### Removed
@@ -20,6 +18,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 ### Security -->
+
+## Unreleased - 2022-XX-XX
+
+### Changed
+
+- Updated dependencies (including `packable`);
 
 ## 1.0.0-beta.4 - 2022-08-15